Fix recursion error.
Close values were causing the recursion limit to be exceeded. This was
caused by incorrect parsing of the binary index while splitting. For
instance, `bin(x)[2:]` would be `"1"` instead of `"001"` for octrees.

In addition, an unknown bug was causing the value for the centre to be
rounded down.

Both problems were fixed, and n_dims was made implicit from the number
of coordinates in the centre parameter.