Switch y and z axes when inputting poses.
The motion capture pose has x as lateral, y as longitudinal, and z as vertical. However, OpenGL displays x as lateral, y as vertical, and z as longitudinal.

Furthermore, forward is positive in ROS's pose, but negative for OpenGL because it works with respect to the origin, so we switch the sign there.